Updated 4kvideodownloader app version

+++ ubuntu.sh
@@ -4,7 +4,7 @@ system_release="`lsb_release -sr`"
 system_architecture="`uname -m`"
 
 echo "LINUX PERSONAL SCRIPT (UBUNTU)"
-echo "Version: 2023.2.28-1010"
+echo "Version: 2023.3.2-1050"
 echo "Author: Danilo Ancilotto"
 echo "System: $system"
 echo "Architecture: $system_architecture"
@@ -42,7 +42,7 @@ dpkgInstall() {
 menuConf() {
   source_file="/usr/share/applications/$2"
   target_file="$1/$2"
-  if [ -f "$source_file" ]
+  if [ -f "$source_file" ] && [ "$5" != "--no-replace-file" ]
   then
     cp -fv "$source_file" "$target_file"
   fi
@@ -159,7 +159,7 @@ printLine "4K Video Downloader"
 root_app_name="4kvideodownloader"
 root_app_subdir="$root_app_dir/$root_app_name"
 root_app_cversion="`sudo cat "$root_app_subdir/version.txt"`"
-root_app_version="4.23.1"
+root_app_version="4.23.2"
 
 if [ "$root_app_cversion" != "$root_app_version" ]
 then
